Use framework prefilter for massive perf improvement when running fraction of tests
Implements creation of a prefilter only when VS is passing in a list of test cases. As OsirisTerje is already working on prefiltering with the TfsTestFilter, I didn't work on implementing anything else with preflighting in that method. VS seems to pass in a list of test cases when you are running a single test, which does not pass in a TfsTestFilter, and therefore, shouldn't affect his work.