Remove patching NUnit internal logger
We couldn't reproduce the issue we identified when we started to work
on adding support for NUnit on our new platform.
After some tests with framework v3 and v4 in command line and in VS,
I wasn't able to reproduce any issue so I am removing the patching code.

Note that the field was renamed in v4 (starts with some underscore)
which is causing the code to fail on line 106